refactor: パッケージ名にアンダーバーが含まれていたため削除した。
