--- app/src/main/kotlin/jp/co/yumemi/android/code_check/MainApplication.kt
@@ -1,15 +1,0 @@
-package jp.co.yumemi.android.code_check
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-import timber.log.Timber
-
-@HiltAndroidApp
-class MainApplication : Application() {
-    override fun onCreate() {
-        super.onCreate()
-        if (BuildConfig.DEBUG) {
-            Timber.plant(Timber.DebugTree())
-        }
-    }
-}
