--- app/src/main/kotlin/jp/co/yumemi/android/code_check/TopActivity.kt
@@ -1,10 +1,0 @@
-/*
- * Copyright © 2021 YUMEMI Inc. All rights reserved.
- */
-package jp.co.yumemi.android.code_check
-
-import androidx.appcompat.app.AppCompatActivity
-import dagger.hilt.android.AndroidEntryPoint
-
-@AndroidEntryPoint
-class TopActivity : AppCompatActivity(R.layout.activity_top)
