+++ app/src/main/kotlin/jp/co/yumemi/android/codecheck/MainApplication.kt
@@ -1,0 +1,15 @@
+package jp.co.yumemi.android.codecheck
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+import timber.log.Timber
+
+@HiltAndroidApp
+class MainApplication : Application() {
+    override fun onCreate() {
+        super.onCreate()
+        if (BuildConfig.DEBUG) {
+            Timber.plant(Timber.DebugTree())
+        }
+    }
+}
