+++ app/src/main/kotlin/jp/co/yumemi/android/codecheck/TopActivity.kt
@@ -1,0 +1,10 @@
+/*
+ * Copyright © 2021 YUMEMI Inc. All rights reserved.
+ */
+package jp.co.yumemi.android.codecheck
+
+import androidx.appcompat.app.AppCompatActivity
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class TopActivity : AppCompatActivity(R.layout.activity_top)
