+++ app/src/main/res/navigation/nav_graph.xml
@@ -6,7 +6,7 @@
 
     <fragment
         android:id="@+id/repositorySearchFragment"
-        android:name="jp.co.yumemi.android.code_check.ui.search.RepositorySearchFragment"
+        android:name="jp.co.yumemi.android.codecheck.ui.search.RepositorySearchFragment"
         android:label="@string/app_name">
         <action
             android:id="@+id/action_repositorySearchFragment_to_repositoryDetailFragment"
@@ -15,11 +15,11 @@
 
     <fragment
         android:id="@+id/repositoryDetailFragment"
-        android:name="jp.co.yumemi.android.code_check.ui.detail.RepositoryDetailFragment"
+        android:name="jp.co.yumemi.android.codecheck.ui.detail.RepositoryDetailFragment"
         android:label="@string/app_name">
         <argument
             android:name="item"
-            app:argType="jp.co.yumemi.android.code_check.data.model.RepositoryItem" />
+            app:argType="jp.co.yumemi.android.codecheck.data.model.RepositoryItem" />
     </fragment>
 
 </navigation>
